OpenAPI: Refactor module to `responder.ext.openapi`
It has been `responder.ext.schema` before.
